+++ src/easy_chat.c
@@ -5006,7 +5006,7 @@ static void HideStartSelectButtons(void)
 
 static void TryAddInterviewObjectEvents(void)
 {
-    int graphicsId;
+    u16 graphicsId;
     u8 spriteId;
 
     switch (GetDisplayedPersonType())
